# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CB Civic Narrative API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface NarrativeRequest {
  lat: number;
  lng: number;
  address?: string;
}

export interface PropertyData {
  address?: string;
  borough?: string;
  yearBuilt?: number;
  landUse?: string;
  zoning?: string;
  bbl?: string;
  lotArea?: number;
  bldgArea?: number;
  numFloors?: number;
  unitsTotal?: number;
  ownerName?: string;
}

export interface LandmarkData {
  isLandmark?: boolean;
  landmarkName?: string;
  designation?: string;
  designationDate?: string;
  borough?: string;
}

export interface MortgageData {
  zipCode?: string;
  denialRate?: number;
  cityAvgDenialRate?: number;
  totalApplications?: number;
  totalDenials?: number;
  year?: number;
}

export interface PPPData {
  zipCode?: string;
  totalLoans?: number;
  totalAmount?: number;
  avgLoanAmount?: number;
  coverageRate?: number;
  cityAvgCoverageRate?: number;
}

export interface ArchivalPhoto {
  url: string;
  year: number;
  description?: string;
  source?: string;
}

export interface Coordinates {
  lat: number;
  lng: number;
}

export interface NarrativeResponse {
  /** Interleaved markdown narrative from Gemini */
  narrative: string;
  propertyData?: PropertyData;
  landmarkData?: LandmarkData;
  mortgageData?: MortgageData;
  pppData?: PPPData;
  photos?: ArchivalPhoto[];
  address: string;
  coordinates: Coordinates;
}

export interface VisionRequest {
  /** Base64-encoded image data */
  imageBase64: string;
  /** Image MIME type (e.g. image/jpeg, image/png) */
  mimeType?: string;
}

export interface VisionResponse {
  /** Gemini's visual analysis of the building/block */
  visualDescription: string;
}

export interface DemoBlock {
  name: string;
  description: string;
  lat: number;
  lng: number;
  zipCode: string;
}

export interface ErrorResponse {
  error: string;
}