# lib/api-zod/src/generated/types/visionRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CB Civic Narrative API
 * OpenAPI spec version: 0.1.0
 */

export interface VisionRequest {
  /** Base64-encoded image data */
  imageBase64: string;
  /** Image MIME type (e.g. image/jpeg, image/png) */
  mimeType?: string;
}