# lib/api-zod/src/generated/types/visionResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CB Civic Narrative API
 * OpenAPI spec version: 0.1.0
 */

export interface VisionResponse {
  /** Gemini's visual analysis of the building/block */
  visualDescription: string;
}